services/metadata-agent/requirements.txt
# Metadata Agent - Python Dependencies

# Web framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0

# Image metadata extraction
pillow>=10.1.0
exifread>=3.0.0

# Video metadata (optional - only if MediaInfo installed on system)
# pymediainfo>=6.1.0

# HTTP client
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6
python-dateutil>=2.8.2
